requirements.txt
pytest
pytest-base-url
beautifulsoup4
pandas
numpy
matplotlib
cython
wordcloud
pyspark
nltk
python-dotenv
tweepy
pysql
ipykernel